chore(exec source): split tests into own module file
